Update spring to 5.x version , update elasticsearch to 5.5. Remove Dojo.

+++ infovip-web/src/main/java/com/github/infovip/core/web/response/datatable/DefaultDataTable.java
@@ -56,7 +56,7 @@ public abstract class DefaultDataTable implements IDataTable {
 	}
 	
 	public DataTableResult result() {
-		formatter(em.createQuery(query).setFirstResult(pageable.getOffset()).setMaxResults(pageable.getPageSize()) .getResultList());
+		formatter(em.createQuery(query).setFirstResult( Long.valueOf( pageable.getOffset() ).intValue() ) .setMaxResults(pageable.getPageSize()) .getResultList());
 		return result;
 	}
